[runtime-on-k8s] runtime-on-k8s

+++ adapter/runtime-manager/src/main/java/org/apache/rocketmq/eventbridge/adapter/runtime/manager/worker/WorkerResource.java
@@ -21,6 +21,6 @@ import lombok.Data;
 
 @Data
 public class WorkerResource {
-    Double cpu;
-    Double memory;
+    Long cpu;
+    Long memory;
 }
